identity_repository refactoring
